+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Master
 
+* Rev-default Ruby version to be the latest patch release of last years Ruby version 2.5.3 (https://github.com/heroku/heroku-buildpack-ruby/pull/846)
 * Allow apps to enable `RUBYOPT=--jit` (https://github.com/heroku/heroku-buildpack-ruby/pull/848)
 
 ## v197 (12/18/2018)