+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Master
 
+* Default Ruby version is now 2.5.5, was previously 2.5.3 (https://github.com/heroku/heroku-buildpack-ruby/pull/863)
+
 ## v200 (3/7/2019)
 
 * Fix: Environment variables not being exported for other buildpacks in CI (https://github.com/heroku/heroku-buildpack-ruby/pull/858)